requirements.txt
Flask
google-genai
openai
anthropic
jinja2
pyyaml
python-dotenv
pydantic
reportlab
pypdf
python-docx
